fix: extracting middleware to its own files
[ci skip]
